Added iOS Xcode libary and example projects
Uses the ios-cmake toolchain to create an Xcode project that can be easily imported into another App project.

To re-generate the fineftp-server library Xcode project, just run the `CMakeiOS.sh` script, then open the generated Xcode project located at `xcodeproj/fineftp.xcodeproj`, click on the `fineftp` project in the Project navigator panel on the left, click the `server` target, click the `Build Settings` tab, and change `Architectures` to `Standard architectures` to allow it to build correctly for both iOS devices and Simulator.

TODO: Need to see if there's a way to get the ios-cmake toolchain to automatically set the architecture. It seems like they only have the option to pre-define specific archs rather than using Xcode's default which is best.