Fixed Xcode header search paths to be relative
It looks like the ios-cmake toolchain unfortuntely uses absolute paths for the header search paths, making the generated project non-portable. I manually adjusted the search paths to use the `$(PROJECT_DIR)` variable, so that it will work automatically if used as a git submodule in another project.

When regenerating the project using the `CMakeiOS.sh` script, it will be necessary to do this each time unless I can find a way to have the toolchain do it automatically.