Strip non-ASCII from file names as a temp solution
Without doing this, trying to transfer a file with a non-ASCII character in the file name will fail to open the fstream. Ideally we would want to completely support those file names, but as a temporary fix this at least allows the files to be transferred, even if they end up with a slightly different file name.

Potential issues:
- Will probably fail trying to read or even list such files
- If two files have the same name after removing these chars, there can be conflicts
- File name information is lost